make filter logic a selector in kea
